Add files via upload
Includes: Includes a way to get the coordinates of the map when clicked by an user for my additional function. Trying to make a new class of polygon objects that define each state included by the points in the doc.

Why: It helps for the user to zoom into states if they wish. Also, creating individual states for the use of the other methods helps to do this as well as many other foreseen methods.

This is an update to the previous update to this class. 

Improve: Finishing the methods to do this is going to be priority, After that, cleaning up the code to make this as fast as possible is going to improve it.